[DLMED] add IO folder
