+++ monai/transforms/__init__.py
@@ -14,6 +14,8 @@ from .croppad.array import *
 from .croppad.dictionary import *
 from .intensity.array import *
 from .intensity.dictionary import *
+from .io.array import *
+from .io.dictionary import *
 from .spatial.array import *
 from .spatial.dictionary import *
 from .utility.array import *
